Implement codefix
